// services/auth-service/src/auth/dto/register.dto.d.ts
import { UserRole } from '@omc-erp/shared-types';
export declare class RegisterDto {
    email: string;
    username: string;
    password: string;
    firstName: string;
    lastName: string;
    phoneNumber?: string;
    tenantId: string;
    role?: UserRole;
}
//# sourceMappingURL=register.dto.d.ts.map